Add tailwind dependency and make it look sexy
