+++ index.html
@@ -7,7 +7,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Datoun ADMIN</title>
   </head>
-  <body>
+  <body class="leading-normal tracking-normal bg-slate-50">
     <div id="app"></div>
      <script type="module" src="/src/main.js"></script>
   </body>
