+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "Datoun.cz",
+  "name": "datoun.cz",
   "version": "0.0.1",
   "scripts": {
     "dev": "vite",
@@ -15,6 +15,13 @@
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^3.0.1",
+    "autoprefixer": "^10.4.12",
+    "postcss": "^8.4.16",
+    "tailwindcss": "^3.1.8",
+    "@tailwindcss/aspect-ratio": "^0.4.2",
+    "@tailwindcss/forms": "^0.5.3",
+    "@tailwindcss/line-clamp": "^0.4.2",
+    "@tailwindcss/typography": "^0.5.7",
     "vite": "^3.0.4"
   }
 }
