+++ src/App.vue
@@ -1,164 +1,13 @@
 <script setup>
-import Vyrobna from './components/Vyrobna.vue'
-import BoxAdder from './components/BoxAdder.vue'
-import { add_to_list , remove_from_list } from './components/ListControl.js'
-import { ref } from 'vue'
-import axios from 'axios'
-import swal from 'sweetalert';
-
-// Odesilani JSONu do databaze
-const axios_instance = axios.create({
-  baseURL: 'https://wiki.smid.io/',
-  timeout: 1000,
-  headers: {
-    'Authorization': 'Bearer ' + '8ebe7bf636fde29d64a19dcd144ad569be5791c1e9a2e3caba9809ac3d872d58'
-  }
-});
-
-function submitHandler (inputData){
-  axios_instance.post('/indexes/firmy/documents', inputData)
-  .then(function (response) {
-    console.log(response);
-    swal("Skvělé!", "Firma přidána do databáze.", "success")
-    .then(() => { location.reload() })
-  })
-  .catch(function (error) {
-    console.log(error);
-    swal("Ajaj!", "Firma nebyla přidána!", "error")
-    .then(() => { location.reload() });
-  });
-}
-
-// Pridavani a odebirani vyroben
-let aktualni_vyrobny = ref( [ 1 ] );
-
-let list_vyroben = ref([]);
-
+import './tailwind.css'
+import FormVyrobna from "./components/form/FormFirma.vue";
+import PageWrap from "./components/reusable/PageWrap.vue";
 </script>
 
 <template>
 
-  <h1>Přidat novou firmu</h1>
-
-  <FormKit
-    type="form"
-    submit-label="Poslat"
-    #default="{ value }"
-    @submit="submitHandler"
-    :config="{ validationVisibility: 'dirty' }"
-    incomplete-message="Nějaké pole nebylo vyplněno správně..."
-  >
-    <!-- ID -->
-    <FormKit
-      name="id"
-      type="hidden"
-      :value='Date.now()'
-    />
-    <!-- Jmeno firmy -->
-    <FormKit
-      type="text"
-      label="Jméno firmy"
-      name="jmeno_firmy"
-      placeholder=""
-      help="Přidej pouze firmu, která zatím není v databázi."
-      validation="required|length:2,100"
-      :validation-messages="{
-        required: 'Jméno firmy je povinné pole.',
-        length: 'Jméno musí být dlouhé 2 až 100 znaků.',
-      }"
-    />
-
-    <!-- Pridavani Aliasuu +- -->
-    <BoxAdder
-      name="aliasy"
-      title="Aliasy"
-      description="Značky pod kterými firma vyrábí, aliasy."
-      :minimumBoxes="0"
-    />
-
-    <!-- Popisek firmy -->
-    <FormKit
-      type="textarea"
-      label="Popisek firmy"
-      name="popisek_firmy"
-      rows="8"
-      placeholder=""
-      help="Krátký motivační popisek většinou najdete na stránkách firmy."
-      validation="required|length:10,10000"
-      :validation-messages="{
-        required: 'Popisek firmy je povinné pole.',
-        length: 'Popisek musí být dlouhé 10 až 10000 znaků.',
-      }"
-    />
-
-    <!-- Eshop firmy -->
-    <FormKit
-      type="url"
-      label="E-shop firmy (zkopírujte)"
-      name="eshop"
-      placeholder=""
-      help="Člověk musí mít možnost si od firmy něco koupit (B2C)."
-      validation="required|url"
-      :validation-messages="{
-        required: 'Adresa eshopu firmy je povinné pole.',
-        url: 'Adresa eshopu musí být platná URL.',
-      }"
-    />
-
-    <!-- Je zaniklá? -->
-    <FormKit
-      type="checkbox"
-      label="Firma je zaniklá"
-      help="Zašrkněte pokud firma zkrachovala nebo přestala vyrábět v Česku."
-      name="zanikla"
-      value=false
-      validation="required"
-      :validation-messages="{
-        required: 'Musíte označit jestli je firma zaniklá, nebo ne.',
-      }"
-    />
-
-    <!-- Přidat upload loga firmy -->
-    <!-- TBA -->
-
-    <!-- Pridavani vyroben +- -->
-    <FormKit
-      v-model="list_vyroben"
-      type="list"
-      name="vyrobny"
-    >
-      <Vyrobna v-for="item in aktualni_vyrobny" :key="item" :factory_number="item" />
-    </FormKit>
-    <div class="row">
-      <FormKit
-        type="button"
-        label="Přidat výrobnu"
-        class="column"
-        @click="add_to_list(aktualni_vyrobny)"
-      />
-      <FormKit
-        type="button"
-        label="Odebrat výrobnu"
-        class="column"
-        @click="remove_from_list(aktualni_vyrobny)"
-      />
-    </div>
-
-    <!-- Zobraz JSON pro kontrolu -->
-    <pre wrap>{{ value }}</pre>
-    </FormKit>
+  <PageWrap>
+    <FormVyrobna/>
+  </PageWrap>
 
 </template>
-
-<style>
-h1{
-  font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-}
-.row {
-  display: flex;
-}
-
-.column {
-  flex: 50%;
-}
-</style>
