+++ src/components/reusable/PageWrap.vue
@@ -1,0 +1,11 @@
+<script setup>
+
+</script>
+
+<template>
+
+  <div class="container pt-8 mb-16">
+    <slot />
+  </div>
+
+</template>
