+++ src/components/reusable/TextBox.vue
@@ -1,0 +1,50 @@
+<script setup>
+  import { remove_from_list_by_id } from '../../utils/ListControl.js'
+
+  defineProps({
+    number: {
+      type: Number,
+      required: true
+    },
+    arrayOfTextfields: {
+      type: Array,
+      required:true
+    },
+    minimum:{
+      type: Number,
+      required:true
+    },
+    removable: {
+      type: Boolean,
+      required: true
+    }
+  })
+</script>
+
+<template>
+<!-- Jaké kategorie výrobků se ve výrobně vyrábí -->
+  <div class="flex gap-4">
+    <FormKit
+        type="text"
+        label=""
+        help=""
+        :placeholder=number
+        validation="required|length:2,1000"
+        outer-class="flex-grow"
+        :validation-messages="{
+        required: 'Povinné pole.',
+        length: 'Délka musí být mezi 2 a 1000 znaků.',
+    }"
+    />
+    <!-- ➖ -->
+    <FormKit
+        type="button"
+        label="➖"
+        class="column"
+        style="background-color:pink;"
+        :disabled="!removable"
+        @click="remove_from_list_by_id(arrayOfTextfields, number, minimum)"
+    />
+  </div>
+
+</template>
